Change MovieScreen.kt
