--- features/movie/src/main/java/com/andhiratobing/moviecompose/movie/MovieScreen.kt
@@ -1,35 +1,0 @@
-package com.andhiratobing.moviecompose.movie
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.sp
-
-@Composable
-fun MovieScreen() {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .wrapContentSize(Alignment.Center)
-    ) {
-        Text(
-            text = " Movie Screen",
-            color = Color.Black,
-            fontSize = 25.sp,
-            modifier = Modifier.align(Alignment.CenterHorizontally)
-        )
-    }
-}
-
-
-@Preview(showBackground = true)
-@Composable
-fun MovieScreenPreview() {
-    MovieScreen()
-}
